html/rendering/non-replaced-elements/tables/table-th-align-default.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Table th default align</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/rendering.html#tables-2">
<link rel="author" title="Intel" href="http://www.intel.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
  table {
    width: 200px;
  }
</style>

<table>
  <thead>
    <tr>
      <th id="th">Table th default align is center</th>
      <th>Value</th>
    </tr>
  </thead>
</table>

<script>

test(() => {
  let elem = document.getElementById("th");
  let css = window.getComputedStyle(elem, null).getPropertyValue("text-align");
  assert_equals(css, "center");
}, `th default align attribute is center`)

</script>
